LOF with full display shows links to file/fid #2033
Originally I treated this as an exception.  All the content is in the list, so there's nothing to be gained by having a link to the file page which has no useful content.  Except that the file page is useful to link to.  Someone wanting to link to one of many files needs that link.  Now title/download is on all the modes - no more exception for full view mode.