LoF/LoP ajax pager #2033
The pager js didn't make it into the new sv list or that commit got lost.  Also had to be updated to work with LOF.

I tested LOF.  That it paged through content correclty was no surprise.  That video thumbnails and flash vids worked after ajax loads was a surprise but I'm not complaining.