Video indicator hover change #2033
List of videos in teaser display had separate links for title and thumbnail.  This wasn't very obvious.  Now the thumbnail link differentiates itself by changing the video-indicator icon when its hovered over.  Hovering on the title link does not change the icon.  This should make it apparent the links aren't the same.