Don't show past repeat events in upcoming sv list #4888
The LOP selects events with an EFQ.  It only selects events in the future.

Repeated events are a single node with many date fields.  To show them separated out in the LOP, we explode the node into many clones, each identical but for the date field.

The combination of these two factors allowed events with future and past dates to be selected, then treated the past dates as displayable nodes.  This commit prevents them from being cloned.