SV list selects singular subsites #4911
This actually took less code than the all or nothing approach.  What took a little while was writing the update hook to convert the old option into the new one.  There were a few different iterations of what the old option looked like (sometimes it was a string, sometimes an array, sometimes it was empty) and I had to take those all into account.