Began according/toggle/slider view for lop #4857
This is the generic approach.  It makes any node slidable.  We'll later make a whitelist of supported content types.

(The other approach is to make a view mode for each content type.  These view modes will automagically appear in the LOP.  I can't start that till I have a list of content types though.)

All the changes are contained in the slider.inc plugin, except for a hook_preprocess_node.