LOP grid WIP #4959
Ditched one approach from LOF grid.  It's apparent that we'll be using grid thumb/teaser/full/title/etc.  Combining the selects no longer makes sense.  Dropped that.  Now am trying to have a layout select for choosing grid or list.

I would like to avoid making new view modes and then sorting them based on whether or not 'grid' appears in the title.  That seems too verbose.  It basically doubles the number of view modes and the length of os_files_file_view().  Using the view modes we have and changing them when grid is selected seems a bit more sane.  But that function will get ugly fast if each of the file type's view modes also has a case for list and grid.  Sooo....

Instead the current approach is going to be to let the LOF get build arrays for each file as usual.  In grid mode, alter them.  This means:
1. Change full video to use a thumbnail
2. Give title modes a thumbnail (or introduce a thumbnail mode)
3. Give images a different style.  The os_files_* ones resize width, but keep height proportional.  Grid needs heights to be consistent, unless we're going to do the pinterest thing.